LayoutTests/media/modern-media-controls/ios-inline-media-controls/ios-inline-media-controls-constructor.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>IOSInlineMediaControls</code> constructor.");

const mediaControls = new IOSInlineMediaControls;

shouldBeEqualToString("mediaControls.element.localName", "div");
shouldBeEqualToString("mediaControls.element.className", "media-controls ios inline");
shouldBe("mediaControls.layoutTraits", "LayoutTraits.iOS");

debug("");
debug("Controls bar");
shouldBe("mediaControls.controlsBar.children.length", "2");
shouldBeTrue("mediaControls.controlsBar.children[0] instanceof ButtonsContainer");
shouldBeTrue("mediaControls.controlsBar.children[1] instanceof ButtonsContainer");

debug("");
debug("Left container");
shouldBe("mediaControls.controlsBar.children[0].buttons", "[mediaControls.playPauseButton, mediaControls.skipBackButton]");
shouldBeEqualToString("mediaControls.controlsBar.children[0].element.className", "buttons-container left");
shouldBe("mediaControls.controlsBar.children[0].leftMargin", "24");
shouldBe("mediaControls.controlsBar.children[0].rightMargin", "24");
shouldBe("mediaControls.controlsBar.children[0].buttonMargin", "24");

debug("");
debug("Right container");
shouldBe("mediaControls.controlsBar.children[1].buttons", "[mediaControls.airplayButton, mediaControls.pipButton, mediaControls.fullscreenButton]");
shouldBeEqualToString("mediaControls.controlsBar.children[1].element.className", "buttons-container right");
shouldBe("mediaControls.controlsBar.children[1].leftMargin", "24");
shouldBe("mediaControls.controlsBar.children[1].rightMargin", "24");
shouldBe("mediaControls.controlsBar.children[1].buttonMargin", "24");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
